server/requirements.txt
annotated-types==0.6.0
anyio==4.2.0
boto3==1.34.29
botocore==1.34.29
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
distro==1.9.0
exceptiongroup==1.2.0
fastapi==0.109.0
h11==0.14.0
httpcore==1.0.2
httpx==0.26.0
idna==3.6
jmespath==1.0.1
mangum==0.17.0
openai==1.8.0
pydantic==2.5.3
pydantic_core==2.14.6
python-dateutil==2.8.2
python-dotenv==1.0.0
python-multipart==0.0.6
requests==2.31.0
s3transfer==0.10.0
six==1.16.0
sniffio==1.3.0
starlette==0.35.1
tqdm==4.66.1
typing_extensions==4.9.0
urllib3==1.26.18
uvicorn==0.26.0
